Enforce strict compiler warnings
This exposed a bunch of issues that needed dealing with to ensure the
code is clean and sane. Notably the dlopen/dlsym routine has been altered
to closer match the LSI approach of safe symbol binding, by not attempting
to directly cast the result of a dlsym operation. Instead, if we succeed
in getting the dlsym() pointer, we memcpy this to the target and ensure
we have the correct constraints.

Note that in sanitizing the log helpers, I opted to remove the varargs
ability from FATAL_ERRNO given this is used exactly like perror() and
there are no examples currently using varargs with this in the tree.
This allowed me to keep the log helpers as macros and not have to implement
wrapper functions.

+++ daemon/governors-query.c
@@ -116,12 +116,12 @@ const char *get_gov_state()
 
 		/* Grab the file length */
 		fseek(f, 0, SEEK_END);
-		int length = ftell(f);
+		long length = ftell(f);
 		fseek(f, 0, SEEK_SET);
 
 		char contents[length];
 
-		if (fread(contents, 1, length, f) > 0) {
+		if (fread(contents, 1, (size_t)length, f) > 0) {
 			/* Files have a newline */
 			strtok(contents, "\n");
 			if (strlen(governor) > 0 && strncmp(governor, contents, 64) != 0) {
